Move app locations to app package
