+++ apps/web/package.json
@@ -29,6 +29,7 @@
     "@boxyhq/saml-jackson": "1.18.6",
     "@calcom/app-store": "*",
     "@calcom/app-store-cli": "*",
+    "@calcom/app-store-types": "*",
     "@calcom/core": "*",
     "@calcom/dayjs": "*",
     "@calcom/embed-core": "workspace:*",
