+++ packages/app-store/_types/index.ts
@@ -1,0 +1,1 @@
+export * from "./locations";
