+++ packages/app-store/_types/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@calcom/app-store-types",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./dist/index.js",
+  "types": "./dist/index.js",
+  "scripts": {
+    "build": "tsc --build --force tsconfig.json",
+    "build:watch": "tsc --build --force ./tsconfig.json --watch",
+    "post-install": "yarn build"
+  }
+}
