+++ packages/app-store/utils.ts
@@ -1,9 +1,9 @@
 import type { AppCategories } from "@prisma/client";
 
+import type { EventLocationType } from "@calcom/app-store-types";
 // If you import this file on any app it should produce circular dependency
 // import appStore from "./index";
 import { appStoreMetadata } from "@calcom/app-store/appStoreMetaData";
-import type { EventLocationType } from "@calcom/app-store/locations";
 import logger from "@calcom/lib/logger";
 import { getPiiFreeCredential } from "@calcom/lib/piiFreeData";
 import { safeStringify } from "@calcom/lib/safeStringify";
