+++ packages/features/bookings/lib/handleNewBooking.ts
@@ -9,13 +9,10 @@ import type { Logger } from "tslog";
 import { v5 as uuidv5 } from "uuid";
 import type z from "zod";
 
+import { MeetLocationType, OrganizerDefaultConferencingAppType } from "@calcom/app-store-types";
 import processExternalId from "@calcom/app-store/_utils/calendars/processExternalId";
 import { metadata as GoogleMeetMetadata } from "@calcom/app-store/googlevideo/_metadata";
-import {
-  MeetLocationType,
-  OrganizerDefaultConferencingAppType,
-  getLocationValueForDB,
-} from "@calcom/app-store/locations";
+import { getLocationValueForDB } from "@calcom/app-store/locations";
 import type { EventTypeAppsList } from "@calcom/app-store/utils";
 import { getAppFromSlug } from "@calcom/app-store/utils";
 import EventManager from "@calcom/core/EventManager";
