+++ packages/features/eventtypes/lib/types.ts
@@ -1,6 +1,6 @@
 import type { z } from "zod";
 
-import type { EventLocationType } from "@calcom/core/location";
+import type { EventLocationType } from "@calcom/app-store-types";
 import type { ChildrenEventType } from "@calcom/features/eventtypes/components/ChildrenEventTypeSelect";
 import type { PeriodType, SchedulingType } from "@calcom/prisma/enums";
 import type { BookerLayoutSettings, EventTypeMetaDataSchema } from "@calcom/prisma/zod-utils";
