+++ packages/trpc/server/routers/viewer/bookings/confirm.handler.ts
@@ -1,8 +1,8 @@
 import { Prisma } from "@prisma/client";
 
 import appStore from "@calcom/app-store";
+import type { LocationObject } from "@calcom/app-store-types";
 import { getLocationValueForDB } from "@calcom/app-store/locations";
-import type { LocationObject } from "@calcom/app-store/locations";
 import { sendDeclinedEmails } from "@calcom/emails";
 import { getCalEventResponses } from "@calcom/features/bookings/lib/getCalEventResponses";
 import { handleConfirmation } from "@calcom/features/bookings/lib/handleConfirmation";
