+++ packages/trpc/server/routers/viewer/eventTypes/bulkUpdateToDefaultLocation.handler.ts
@@ -1,4 +1,4 @@
-import type { LocationObject } from "@calcom/app-store/locations";
+import type { LocationObject } from "@calcom/app-store-types";
 import { getAppFromSlug } from "@calcom/app-store/utils";
 import { prisma } from "@calcom/prisma";
 import { userMetadata as userMetadataSchema } from "@calcom/prisma/zod-utils";
